# specs/001-github-repo-research/contracts/research-web.openapi.yaml
openapi: 3.1.0
info:
  title: GitHub Repository Research API
  version: 0.1.0
  description: |
    OpenAPI contract for the research job submission and status polling APIs.
    The primary human-readable output is a structured Markdown report, while
    these endpoints expose machine-readable JSON contracts.
servers:
  - url: http://127.0.0.1:8000
    description: Local development server

tags:
  - name: research
    description: Repository research job lifecycle

paths:
  /api/v1/research/jobs:
    post:
      tags: [research]
      summary: Submit a repository research job
      operationId: submitResearchJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchRequest'
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
        '400':
          description: Invalid repository URL or unsupported input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limited
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal error during acceptance flow
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/v1/research/jobs/{job_id}:
    get:
      tags: [research]
      summary: Poll research job status
      operationId: getResearchJobStatus
      parameters:
        - in: path
          name: job_id
          required: true
          description: Research job identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current job state snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollingStatusResponse'
        '404':
          description: Job not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal error while loading job state
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

components:
  schemas:
    ResearchRequest:
      type: object
      additionalProperties: false
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Public GitHub repository URL in https://github.com/{owner}/{repo} format
          examples:
            - https://github.com/tiangolo/fastapi
        options:
          type: object
          description: Optional analysis tuning flags
          additionalProperties: true

    JobAcceptedResponse:
      type: object
      additionalProperties: false
      required: [job_id, status]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending]

    PollingStatusResponse:
      type: object
      additionalProperties: false
      required: [job_id, status, progress_msg, updated_at]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress_msg:
          type: string
        updated_at:
          type: string
          format: date-time
        instant_data:
          $ref: '#/components/schemas/InstantMetadata'
        report:
          $ref: '#/components/schemas/ResearchReport'
        ai_fallback:
          type: boolean
        ai_fallback_reason:
          type: string
        error:
          type: string

    InstantMetadata:
      type: object
      additionalProperties: false
      required: [name, stars, forks]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        stars:
          type: integer
          minimum: 0
        forks:
          type: integer
          minimum: 0
        primary_lang:
          type: string
          nullable: true
        last_commit_at:
          type: string
          format: date-time
          nullable: true

    StructuredAIResponse:
      type: object
      additionalProperties: false
      required:
        - project_purpose
        - tech_stack
        - architecture
        - risk_level
        - ai_confidence
      properties:
        project_purpose:
          type: string
          description: Brief project purpose summary
        tech_stack:
          type: array
          items:
            type: string
        architecture:
          type: string
        risk_level:
          type: string
          enum: [Low, Medium, High]
        ai_confidence:
          type: string
          enum: [High, Low]
        limitations:
          type: array
          items:
            type: string

    ResearchReport:
      type: object
      additionalProperties: false
      required: [job_id, repository_url, deterministic_data, markdown_report, created_at]
      properties:
        job_id:
          type: string
          format: uuid
        repository_url:
          type: string
          format: uri
        deterministic_data:
          $ref: '#/components/schemas/InstantMetadata'
        ai_insights:
          $ref: '#/components/schemas/StructuredAIResponse'
        ai_fallback_reason:
          type: string
        markdown_report:
          type: string
          description: Structured Markdown final report (primary human-readable artifact)
        created_at:
          type: string
          format: date-time

    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference